requirements.txt
tensorflow==2.16.1
gradio==4.41.0
numpy
Pillow

